+++ src/main/scala/microsites/MicrositeKeys.scala
@@ -16,12 +16,18 @@
 
 package microsites
 
-import com.typesafe.sbt.sbtghpages.GhpagesPlugin.autoImport.ghpagesPushSite
+import com.typesafe.sbt.sbtghpages.GhpagesPlugin.autoImport.{
+  ghpagesBranch,
+  ghpagesNoJekyll,
+  ghpagesPushSite
+}
 import com.typesafe.sbt.site.SitePlugin.autoImport.{makeSite, siteDirectory}
 import com.typesafe.sbt.site.jekyll.JekyllPlugin.autoImport._
 import microsites.util.MicrositeHelper
 import sbt.Keys._
 import sbt._
+import sbt.complete.DefaultParsers.OptNotSpace
+import sbtorgpolicies.github.GitHubOps
 import tut.Plugin._
 
 trait MicrositeKeys {
@@ -32,6 +38,10 @@ trait MicrositeKeys {
   final case object Bitbucket                               extends GitHostingService("Bitbucket")
   final case class Other(value: String)                     extends GitHostingService(value)
 
+  sealed trait PushWith
+  final case object GHPagesPlugin extends PushWith
+  final case object GitHub4s      extends PushWith
+
   object GitHostingService {
     implicit def string2GitHostingService(name: String): GitHostingService = {
       List(GitHub, GitLab, Bitbucket)
@@ -40,62 +50,71 @@ trait MicrositeKeys {
     }
   }
 
-  val makeMicrosite = taskKey[Unit]("Main Task to build a Microsite")
-  val publishMicrosite =
-    taskKey[Unit]("Publish the microsite (using the pushSite task) after build it")
-  val microsite = taskKey[Seq[File]]("Create microsite files")
-  val micrositeConfig =
+  val makeMicrosite: TaskKey[Unit]  = taskKey[Unit]("Main Task to build a Microsite")
+  val microsite: TaskKey[Seq[File]] = taskKey[Seq[File]]("Create microsite files")
+  val micrositeConfig: TaskKey[Unit] =
     taskKey[Unit]("Copy microsite config to the site folder")
-  val micrositeName                 = settingKey[String]("Microsite name")
-  val micrositeDescription          = settingKey[String]("Microsite description")
-  val micrositeAuthor               = settingKey[String]("Microsite author")
-  val micrositeHomepage             = settingKey[String]("Microsite homepage")
-  val micrositeOrganizationHomepage = settingKey[String]("Microsite organisation homepage")
-  val micrositeTwitter              = settingKey[String]("Microsite twitter")
-  val micrositeBaseUrl              = settingKey[String]("Microsite site base url")
-  val micrositeDocumentationUrl =
+  val micrositeName: SettingKey[String]        = settingKey[String]("Microsite name")
+  val micrositeDescription: SettingKey[String] = settingKey[String]("Microsite description")
+  val micrositeAuthor: SettingKey[String]      = settingKey[String]("Microsite author")
+  val micrositeHomepage: SettingKey[String]    = settingKey[String]("Microsite homepage")
+  val micrositeOrganizationHomepage: SettingKey[String] =
+    settingKey[String]("Microsite organisation homepage")
+  val micrositeTwitter: SettingKey[String] = settingKey[String]("Microsite twitter")
+  val micrositeBaseUrl: SettingKey[String] = settingKey[String]("Microsite site base url")
+  val micrositeDocumentationUrl: SettingKey[String] =
     settingKey[String]("Microsite site documentation url")
-  val micrositeHighlightTheme = settingKey[String]("Microsite Highlight Theme")
-  val micrositeConfigYaml =
+  val micrositeHighlightTheme: SettingKey[String] = settingKey[String]("Microsite Highlight Theme")
+  val micrositeConfigYaml: SettingKey[ConfigYml] =
     settingKey[ConfigYml]("Microsite _config.yml file configuration.")
-  val micrositeImgDirectory = settingKey[File](
+  val micrositeImgDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite images directory. By default, it'll be the resourcesDirectory + '/microsite/img'")
-  val micrositeCssDirectory = settingKey[File](
+  val micrositeCssDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite CSS directory. By default, it'll be the resourcesDirectory + '/microsite/css'")
-  val micrositeJsDirectory = settingKey[File](
+  val micrositeJsDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite Javascript directory. By default, it'll be the resourcesDirectory + '/microsite/js'")
-  val micrositeCDNDirectives = settingKey[CdnDirectives](
+  val micrositeCDNDirectives: SettingKey[CdnDirectives] = settingKey[CdnDirectives](
     "Optional. Microsite CDN directives lists (for css and js imports). By default, both lists are empty.")
-  val micrositeExternalLayoutsDirectory = settingKey[File](
+  val micrositeExternalLayoutsDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite external layouts directory. By default, it'll be the resourcesDirectory + '/microsite/layout'")
-  val micrositeExternalIncludesDirectory = settingKey[File](
+  val micrositeExternalIncludesDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite external includes (partial layouts) directory. By default, it'll be the resourcesDirectory + '/microsite/includes'")
-  val micrositeDataDirectory = settingKey[File](
+  val micrositeDataDirectory: SettingKey[File] = settingKey[File](
     "Optional. Microsite Data directory, useful to define the microsite data files " +
       "(https://jekyllrb.com/docs/datafiles/). By default, it'll be the resourcesDirectory + '/microsite/data'")
-  val micrositeExtraMdFiles = settingKey[Map[File, ExtraMdFileConfig]](
-    "Optional. This key is useful when you want to include automatically markdown documents as a part of your microsite, and these files are located in different places from the tutSourceDirectory. The map key is related with the source file, the map value corresponds with the target relative file path and the document meta-information configuration. By default, the map is empty.")
-  val micrositePalette = settingKey[Map[String, String]]("Microsite palette")
-  val micrositeFavicons = settingKey[Seq[MicrositeFavicon]](
+  val micrositeExtraMdFiles: SettingKey[Map[File, ExtraMdFileConfig]] =
+    settingKey[Map[File, ExtraMdFileConfig]](
+      "Optional. This key is useful when you want to include automatically markdown documents as a part of your microsite, and these files are located in different places from the tutSourceDirectory. The map key is related with the source file, the map value corresponds with the target relative file path and the document meta-information configuration. By default, the map is empty.")
+  val micrositePalette: SettingKey[Map[String, String]] =
+    settingKey[Map[String, String]]("Microsite palette")
+  val micrositeFavicons: SettingKey[Seq[MicrositeFavicon]] = settingKey[Seq[MicrositeFavicon]](
     "Optional. List of filenames and sizes for the PNG/ICO files to be used as favicon for the generated site, located in '/microsite/img'. The sizes should be described with a string (i.e.: \"16x16\"). By default, favicons with different sizes will be generated from the navbar_brand2x.jpg file.")
-  val micrositeGithubOwner = settingKey[String]("Microsite Github owner")
-  val micrositeGithubRepo  = settingKey[String]("Microsite Github repo")
-  val micrositeKazariEvaluatorUrl = settingKey[String](
+  val micrositeGithubOwner: SettingKey[String] = settingKey[String]("Microsite Github owner")
+  val micrositeGithubRepo: SettingKey[String]  = settingKey[String]("Microsite Github repo")
+  val micrositeGithubToken: SettingKey[Option[String]] =
+    settingKey[Option[String]]("Microsite Github token for pushing the microsite")
+  val micrositeKazariEvaluatorUrl: SettingKey[String] = settingKey[String](
     "URL of the remote Scala Evaluator to be used by Kazari. Required for Kazari to work. Default: https://scala-evaluator-212.herokuapp.com")
-  val micrositeKazariEvaluatorToken = settingKey[String](
+  val micrositeKazariEvaluatorToken: SettingKey[String] = settingKey[String](
     "Remote Scala Evaluator token to be used by Kazari. Required for Kazari to work. Default: token compatible with the Scala Exercises remote evaluator.")
-  val micrositeKazariGithubToken = settingKey[String](
+  val micrositeKazariGithubToken: SettingKey[String] = settingKey[String](
     "GitHub token to be used by Kazari. Required for Kazari to perform certain actions (i.e. save Gists). Default: empty string")
-  val micrositeKazariCodeMirrorTheme = settingKey[String](
+  val micrositeKazariCodeMirrorTheme: SettingKey[String] = settingKey[String](
     "Optional. CodeMirror theme to be used by Kazari in its modal editor. Default: monokai")
-  val micrositeKazariDependencies = settingKey[Seq[KazariDependency]](
-    "Optional. List of dependencies needed to compile the code to be evaluated by Kazari (set of groupId, artifactId, scalaVersion and versionId). Default: empty list")
-  val micrositeKazariResolvers = settingKey[Seq[String]](
+  val micrositeKazariDependencies: SettingKey[Seq[KazariDependency]] =
+    settingKey[Seq[KazariDependency]](
+      "Optional. List of dependencies needed to compile the code to be evaluated by Kazari (set of groupId, artifactId, scalaVersion and versionId). Default: empty list")
+  val micrositeKazariResolvers: SettingKey[Seq[String]] = settingKey[Seq[String]](
     "Optional. List of resolver urls needed for the provided dependencies to be fetched by Kazari. Default: empty list")
-  val micrositeGitHostingService =
+  val micrositeGitHostingService: SettingKey[GitHostingService] =
     settingKey[GitHostingService]("Service used for git hosting. By default, it'll be GitHub.")
-  val micrositeGitHostingUrl = settingKey[String](
+  val micrositeGitHostingUrl: SettingKey[String] = settingKey[String](
     "In the case where your project isn't hosted on Github, use this setting to point users to git host (e.g. 'https://internal.gitlab.com/<user>/<project>').")
+  val micrositePushSiteWith: SettingKey[PushWith] =
+    settingKey[PushWith](
+      "Determines what will be chosen for pushing the site. The options are sbt-ghpages plugin and github4s library.")
+
+  val publishMicrositeCommandKey: String = "publishMicrosite"
   val micrositeAnalytics = settingKey[String](
     "In the case where your project isn't hosted on Github, use this setting to point users to git host (e.g. 'https://internal.gitlab.com/<user>/<project>').")
 }
@@ -193,13 +212,51 @@ trait MicrositeAutoImportSettings extends MicrositeKeys {
         makeSite,
         micrositeConfig
       )
-      .value,
-    publishMicrosite := Def
-      .sequential(
-        clean,
-        makeMicrosite,
-        ghpagesPushSite
-      )
       .value
   )
+
+  val publishMicrositeCommand: Command = Command(publishMicrositeCommandKey)(_ => OptNotSpace) {
+    (st, _) =>
+      val extracted = Project.extract(st)
+
+      val siteDir: Dir                  = extracted.get(target in makeSite)
+      val noJekyll: Boolean             = extracted.get(ghpagesNoJekyll)
+      val branch: String                = extracted.get(ghpagesBranch)
+      val pushSiteWith: PushWith        = extracted.get(micrositePushSiteWith)
+      val gitHosting: GitHostingService = extracted.get(micrositeGitHostingService)
+      val githubOwner: String           = extracted.get(micrositeGithubOwner)
+      val githubRepo: String            = extracted.get(micrositeGithubRepo)
+      val githubToken: Option[String]   = extracted.get(micrositeGithubToken)
+
+      val cleanState: State = extracted.runTask(clean, st)._1
+      val makeState: State  = extracted.runTask(makeMicrosite, cleanState)._1
+
+      (pushSiteWith, gitHosting) match {
+        case (GHPagesPlugin, _) =>
+          val ref = extracted.get(thisProjectRef)
+          extracted.runAggregated[Unit](ghpagesPushSite in Global in ref, makeState)
+        case (GitHub4s, GitHub) if githubToken.nonEmpty =>
+          val commitMessage = sys.env.getOrElse("SBT_GHPAGES_COMMIT_MESSAGE", "updated site")
+          makeState.log.info(
+            s"""Committing files from ${siteDir.getAbsolutePath} into branch '$branch'
+               | * repo: $githubOwner/$githubRepo
+               | * commitMessage: $commitMessage""".stripMargin)
+          val ghOps: GitHubOps = new GitHubOps(githubOwner, githubRepo, githubToken)
+          if (noJekyll) IO.touch(siteDir / ".nojekyll")
+          ghOps.commitDir(branch, commitMessage, siteDir) match {
+            case Right(_) => makeState.log.info("Success committing files")
+            case Left(e) =>
+              makeState.log.error(s"Error committing files")
+              e.printStackTrace()
+          }
+          makeState
+        case (GitHub4s, GitHub) =>
+          makeState.log.error(
+            s"You must provide a GitHub token through the `micrositeGithubToken` setting for pushing with github4s")
+          makeState
+        case (GitHub4s, hosting) =>
+          makeState.log.warn(s"github4s doens't have support for $hosting")
+          makeState
+      }
+  }
 }