+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 04/21/2017 - Version 0.5.4
+
+Release changes:
+
+* Use GitHub api to publish ghpages ([#169](https://github.com/47deg/sbt-microsites/pull/169))
+* Fixes artifact publish ([#170](https://github.com/47deg/sbt-microsites/pull/170))
+
+
 ## 04/11/2017 - Version 0.5.3
 
 Release changes: